New protocol packages   - new ant build rule for protocol.jar so we can version protocol independent of client   - protocol independenty versioned for use by the tunnelserver.
This is all new code so the current client should be unaffected by this change.  When I 
submit round 2 of the CONFIG in cloud, we can remove the redundant classes.  However, 
we will do the tunnel server changes first with this new code then we can ensure backward
compatibility of the tunnelserver.